tile logic consolidation
